backend/pyproject.toml
[project]
name = "pastexam-backend"
version = "0.1.0"
description = "Past Exam API Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.3",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "minio>=7.2.15",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.3",
]
